Full Project

+++ src/main/java/com/example/week7task/service/serviceImplementation/CommentServiceImpl.java
@@ -21,13 +21,7 @@ public class CommentServiceImpl implements CommentService {
     @Autowired
     private PostRepository postRepository;
 
-    /**
-     * CREATE operation on Comment
-     * @param comment
-     * @param postId
-     * @param userId
-     * @return boolean(true for successful creation and false on failure on create)
-     * */
+
     public boolean createComment(Long userId, Long postId, Comment comment){
         boolean result = false;
 
@@ -46,11 +40,7 @@ public class CommentServiceImpl implements CommentService {
         return result;
     }
 
-    /**
-     * GET operation on Comment
-     * @param postId
-     * @return List of comments
-     * */
+
     public List<CommentMapper> getComments(Long postId){
         List<CommentMapper> comments = new ArrayList();
 
@@ -78,50 +68,29 @@ public class CommentServiceImpl implements CommentService {
         return comments;
     }
 
-    /**
-     * CREATE operation on Comment
-     * @param comment
-     * @param postId
-     * @param commentId
-     * @param user
-     * @return boolean(true for successful creation and false on failure on comment update)
-     * */
     public boolean editComment(Long commentId, User user, Long postId, String comment) {
         boolean status = false;
+        Post post = postRepository.findById(postId).get();
 
-        try {
-            Post post = postRepository.findById(postId).get();
+        Comment data = commentRepository.findCommentById(commentId);
 
-            Comment data = commentRepository.findCommentById(commentId);
-
+        if(post != null && data != null){
             data.setComment(comment);
             data.setUser(user);
             data.setPost(post);
             commentRepository.save(data);
 
             status = true;
-
-        }catch (Exception e) {
-            e.printStackTrace();
         }
-
         return status;
     }
 
-    /**
-     * DELETE operation on Comment
-     * @param commentId
-     * @return boolean(true for successful deletion and false on failure to delete)
-     * */
+
     public boolean deleteComment(Long commentId){
         boolean status =  false;
-
-        try {
-            commentRepository.deleteCommentById(commentId);
-            status = true;
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
+       if(commentRepository.existsById(commentId)){
+           commentRepository.deleteCommentById(commentId);
+       }
         return status;
     }
 }
